Extract KVEventsConfig and KVTransferConfig from configs
